mobadra/assi_big_data_&_airflow/requirements.txt
streamlit
pandas
psycopg2-binary # مكتبة Python للاتصال بـ Postgres